Make board public
Made board, bitboard, square and pieces modules public.
